move IconOutline from harbor header to a separate file
